+++ mail-batch/src/test/java/maeilbatch/mail/weekly/WeeklyMailSenderTest.java
@@ -1,11 +1,13 @@
-package maeilmail.bulksend.sender;
+package maeilbatch.mail.weekly;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.Mockito.mock;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
+import maeilbatch.support.IntegrationTestSupport;
 import maeilmail.mail.MimeMessageCustomizer;
 import maeilmail.question.Question;
 import maeilmail.question.QuestionCategory;
@@ -15,13 +17,12 @@ import maeilmail.subscribe.command.domain.SubscribeFrequency;
 import maeilmail.subscribe.command.domain.SubscribeQuestion;
 import maeilmail.subscribe.command.domain.SubscribeQuestionRepository;
 import maeilmail.subscribe.command.domain.SubscribeRepository;
-import maeilmail.support.IntegrationTestSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.javamail.JavaMailSender;
 
-class QuestionSenderTest extends IntegrationTestSupport {
+class WeeklyMailSenderTest extends IntegrationTestSupport {
 
     @Autowired
     private SubscribeQuestionRepository subscribeQuestionRepository;
@@ -35,16 +36,17 @@ class QuestionSenderTest extends IntegrationTestSupport {
     @Test
     @DisplayName("메일 전송에 성공하면 내역을 생성한다.")
     void handleSuccess() {
+        int expectedQuestionsSize = 5;
         LocalDateTime createdAt = LocalDateTime.of(2025, 5, 1, 7, 0, 0);
         setJpaAuditingTime(createdAt);
-        QuestionSender questionSender = createQuestionSender();
-        SubscribeQuestionMessage message = createMessage(createSubscribe(), createQuestion());
+        WeeklyMailSender weeklyMailSender = createWeeklyMailSender();
+        WeeklyMailMessage message = createMessage(createSubscribe(), createQuestions(expectedQuestionsSize));
 
-        questionSender.handleSuccess(message);
+        weeklyMailSender.handleSuccess(message);
 
         List<SubscribeQuestion> result = subscribeQuestionRepository.findAll();
         assertAll(
-                () -> assertThat(result).hasSize(1),
+                () -> assertThat(result).hasSize(expectedQuestionsSize),
                 () -> assertThat(result.get(0).getCreatedAt()).isEqualTo(createdAt)
         );
     }
@@ -52,30 +54,33 @@ class QuestionSenderTest extends IntegrationTestSupport {
     @Test
     @DisplayName("이미 전송된 질문지라면, 기존 내역을 제거하고 신규 내역을 생성한다.")
     void handleSuccessAlreadySend() {
+        int expectedQuestionsSize = 5;
         setJpaAuditingTime(LocalDateTime.of(2025, 5, 1, 7, 0, 0));
-        QuestionSender questionSender = createQuestionSender();
+        WeeklyMailSender weeklyMailSender = createWeeklyMailSender();
         Subscribe subscribe = createSubscribe();
-        Question question = createQuestion();
-        SubscribeQuestion alreadySendSubscribeQuestion = SubscribeQuestion.success(subscribe, question);
+        List<Question> questions = createQuestions(expectedQuestionsSize);
+        SubscribeQuestion alreadySendSubscribeQuestion = SubscribeQuestion.success(subscribe, questions.get(0));
         subscribeQuestionRepository.save(alreadySendSubscribeQuestion);
         LocalDateTime newCreatedAt = LocalDateTime.of(2025, 5, 2, 7, 0, 0);
         setJpaAuditingTime(newCreatedAt);
 
-        SubscribeQuestionMessage message = createMessage(subscribe, question);
-        questionSender.handleSuccess(message);
+        WeeklyMailMessage message = createMessage(subscribe, questions);
+        weeklyMailSender.handleSuccess(message);
 
         List<SubscribeQuestion> result = subscribeQuestionRepository.findAll();
         assertAll(
-                () -> assertThat(result).hasSize(1),
-                () -> assertThat(result.get(0).getCreatedAt()).isEqualTo(newCreatedAt)
+                () -> assertThat(result).hasSize(expectedQuestionsSize),
+                () -> assertThat(result)
+                        .map(SubscribeQuestion::getCreatedAt)
+                        .allMatch(it -> it.equals(newCreatedAt))
         );
     }
 
-    private QuestionSender createQuestionSender() {
+    private WeeklyMailSender createWeeklyMailSender() {
         JavaMailSender mailSender = mock(JavaMailSender.class);
         MimeMessageCustomizer customizer = mock(MimeMessageCustomizer.class);
 
-        return new QuestionSender(mailSender, customizer, subscribeQuestionRepository);
+        return new WeeklyMailSender(mailSender, customizer, subscribeQuestionRepository);
     }
 
     private Subscribe createSubscribe() {
@@ -84,13 +89,18 @@ class QuestionSenderTest extends IntegrationTestSupport {
         return subscribeRepository.save(subscribe);
     }
 
-    private Question createQuestion() {
-        Question question = new Question("test1", "content", QuestionCategory.BACKEND);
+    private List<Question> createQuestions(int size) {
+        List<Question> questions = new ArrayList<>();
 
-        return questionRepository.save(question);
+        for (int i = 0; i < size; i++) {
+            Question question = new Question("test" + i, "content", QuestionCategory.BACKEND);
+            questions.add(question);
+        }
+
+        return questionRepository.saveAll(questions);
     }
 
-    private SubscribeQuestionMessage createMessage(Subscribe subscribe, Question question) {
-        return new SubscribeQuestionMessage(subscribe, question, "subject", "text");
+    private WeeklyMailMessage createMessage(Subscribe subscribe, List<Question> questions) {
+        return new WeeklyMailMessage(subscribe, questions, "subject", "text");
     }
 }
